Move state, rename view model, and remove unusd providers
