+++ connect-example/src/main/java/com/stripe/android/connect/example/ui/embeddedcomponentmanagerloader/EmbeddedComponentManagerLoaderState.kt
@@ -1,0 +1,11 @@
+package com.stripe.android.connect.example.ui.embeddedcomponentmanagerloader
+
+import com.stripe.android.connect.EmbeddedComponentManager
+import com.stripe.android.connect.PrivateBetaConnectSDK
+import com.stripe.android.connect.example.ui.common.Async
+import com.stripe.android.connect.example.ui.common.Uninitialized
+
+@OptIn(PrivateBetaConnectSDK::class)
+data class EmbeddedComponentManagerLoaderState(
+    val embeddedComponentAsync: Async<EmbeddedComponentManager> = Uninitialized,
+)
